requirements.txt
pytest~=7.0.0
python-dotenv~=0.19.2
selenium~=4.1.0
windows-curses~=2.3.0